[Refactor] Refactor file structure
