+++ src/components/App.test.jsx
@@ -6,9 +6,9 @@ import { render } from '@testing-library/react';
 
 import App from './App';
 
-import regions from '../fixtures/regions';
-import categories from '../fixtures/categories';
-import restaurants from '../fixtures/restaurants';
+import regions from '../../fixtures/regions';
+import categories from '../../fixtures/categories';
+import restaurants from '../../fixtures/restaurants';
 
 jest.mock('react-redux');
