+++ src/components/categories/CategoriesContainer.test.jsx
@@ -6,7 +6,7 @@ import { render, fireEvent } from '@testing-library/react';
 
 import CategoriesContainer from './CategoriesContainer';
 
-import categories from '../../fixtures/categories';
+import categories from '../../../fixtures/categories';
 
 jest.mock('react-redux');
