+++ src/components/categories/CategoryItem.test.jsx
