+++ src/components/regions/Regions.test.jsx
@@ -4,7 +4,7 @@ import { render, fireEvent } from '@testing-library/react';
 
 import Regions from './Regions';
 
-import regions from '../../fixtures/regions';
+import regions from '../../../fixtures/regions';
 
 describe('Regions', () => {
   const handleClickSelectRegion = jest.fn();
