+++ src/components/restaurants/Restaurants.test.jsx
@@ -4,7 +4,7 @@ import { render } from '@testing-library/react';
 
 import Restaurants from './Restaurants';
 
-import restaurants from '../../fixtures/restaurants';
+import restaurants from '../../../fixtures/restaurants';
 
 describe('Restaurants', () => {
   it('renders each item', () => {
