+++ src/components/restaurants/RestaurantsContainer.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 
 import Restaurants from './Restaurants';
-import Loading from '../Loading';
+import Loading from '../common/Loading';
 
 export default function RestaurantsContainer() {
   const { loading, restaurants } = useSelector((({ restaurant }) => ({
