+++ src/index.jsx
@@ -3,9 +3,9 @@ import ReactDOM from 'react-dom';
 
 import { Provider } from 'react-redux';
 
-import store from './redux/store';
+import store from './store';
 
-import App from './App';
+import App from './components/App';
 
 ReactDOM.render(
   <Provider store={store}>
